Created parse metrics to support both backends
